Move ViewBase in prep for splitting UX framework into separate package
